+++ nablaDFT/graphormer/__init__.py
@@ -1,3 +1,2 @@
 from .graphormer_3d import Graphormer3DLightning, Graphormer3D
-from .data_utils import PyGNablaDFT, get_nablaDFT_datasets, get_nablaDFT_test_dataset
 from .schedulers import get_linear_schedule_with_warmup
